requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
peewee==3.17.6
apscheduler==3.10.4
openpyxl==3.2.0b1